Rename facade controllers
